Extract IModelLineProjection into modelLineProjection.ts
